quizgame 5.0

+++ app/src/main/java/me/ppvan/meplace/repository/GameService.kt
@@ -4,7 +4,7 @@ import me.ppvan.meplace.data.GameScore
 import me.ppvan.meplace.data.User
 
 interface GameService {
-    fun getHighestScore(gameName: String): Int?
+    fun getHighestScore(gameName: String): Long?
 
     fun insert(gameScore: GameScore): Unit
 }
